% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cons_plans.R
\name{add_dens_polygon}
\alias{add_dens_polygon}
\title{Add a kernel density polygon}
\usage{
add_dens_polygon(
  x,
  y,
  col,
  lwd = 1.7,
  alpha = c(0.25, 0.75),
  add_pts = FALSE,
  add_poly = TRUE
)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{col}{Colour to add polygon with. Will be made into two levels of
opacity.}

\item{lwd}{lwd Line width}

\item{alpha}{A numeric vector of length 2 that gives the confidence levels
for the two kernel density polygons.}

\item{add_pts}{Logical: should points be added?}

\item{add_poly}{Add polygons?}
}
\description{
Add a kernel density polygon
}
