\name{annotate}
\alias{annotate}
\title{Add annotations to panel}
\usage{
  annotate(label, xfrac = 0.008, yfrac = 0.18, pos = 4,
    cex = 0.9, ...)
}
\arguments{
  \item{label}{The text to add as a label}

  \item{xfrac}{Fraction over from the left}

  \item{yfrac}{Fraction down from the top}

  \item{pos}{Position of text to pass to \code{text()}}

  \item{cex}{cex to pass to \code{text()}}

  \item{...}{Anything else to pass to \code{text()}}
}
\description{
  Add annotations to panel
}

