% Generated by roxygen2 (4.0.0): do not edit by hand
\name{custom_bw}
\alias{custom_bw}
\title{Custom bandwidth}
\usage{
custom_bw(x)
}
\arguments{
  \item{x}{A numeric vector}
}
\description{
Based on \code{bandwidth.nrd} from \pkg{MASS}. This version
takes the absolute value of \code{var} to avoid errors.
}

