% Generated by roxygen2 (4.0.1): do not edit by hand
\name{CVaR}
\alias{CVaR}
\title{Conditional Value at Risk}
\usage{
CVaR(x, probs = 0.05)
}
\arguments{
  \item{x}{A numeric vector}

  \item{probs}{The probability cutoff to pass to the CVaR
  function.}
}
\description{
Get the conditional value at risk.
}

