% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{est_beta_params}
\alias{est_beta_params}
\title{Get beta parameters from mean and variance}
\usage{
est_beta_params(mu, var)
}
\arguments{
\item{mu}{Mean}

\item{var}{Variance}
}
\description{
Get beta parameters from mean and variance
}
