% Generated by roxygen2 (4.0.0): do not edit by hand
\name{get_efficient_frontier}
\alias{get_efficient_frontier}
\title{Get the efficient frontier from mean and variance values}
\usage{
get_efficient_frontier(m, v)
}
\arguments{
  \item{m}{A vector of mean values}

  \item{v}{A vector of variance values}
}
\description{
Get the efficient frontier from mean and variance values
}

