\name{get_port_vals}
\alias{get_port_vals}
\title{Get portfolio mean and variance values from a meta_sim() object}
\usage{
  get_port_vals(x, burn = 1:30)
}
\arguments{
  \item{x}{A list object as returned from \code{meta_sim}}

  \item{burn}{Number of years to throw out at burn in}
}
\description{
  Get portfolio mean and variance values from a meta_sim()
  object
}

