% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cons_plans.R
\name{get_quantile_contour}
\alias{get_quantile_contour}
\title{Get quantile contour}
\usage{
get_quantile_contour(x, alpha = 0.8)
}
\arguments{
\item{x}{Output from \code{\link[MASS]{kde2d}}.}

\item{alpha}{The quantile level.}
}
\description{
Get quantile contour
}
