% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_element}
\alias{is_element}
\title{Check if x is an element of y.}
\usage{
is_element(x, y)
}
\arguments{
\item{x}{An integer to check}

\item{y}{A vector to check if \code{x} is an element of \code{y}.}
}
\description{
Check if x is an element of y.
}
