\name{is_quasi_ext}
\alias{is_quasi_ext}
\title{Return whether there was an instance of quasi extinction}
\usage{
  is_quasi_ext(x, thresh, duration = 1)
}
\arguments{
  \item{x}{A numeric vector.}

  \item{thresh}{The quasi-extinction threshold in absolute
  numbers.}

  \item{duration}{The number of years below the threshold
  before a quasi extinction has occured.}
}
\description{
  Return whether there was an instance of quasi extinction
}

