\name{metafolio-package}
\alias{metafolio-package}
\alias{metafolio}
\docType{package}
\title{
Simulate metapopulations for portfolio optimization
}
\description{
Simulations are based around salmon metapopulations. 
}
\author{
Sean C. Anderson <sean at seananderson.ca>
}
\references{
Portfolio prioritization of salmon conservation. In prep.
}
\keyword{salmon, metapopulations, modern portfolio theory, ecology, conservation}
