\name{monte_carlo_portfolios}
\alias{monte_carlo_portfolios}
\title{Monte Carlo the weights into portfolios and record the simulation}
\usage{
  monte_carlo_portfolios(weights_matrix, n_sims = 500,
    mean_b = 1000, burn = 1:30, ...)
}
\description{
  Monte Carlo the weights into portfolios and record the
  simulation
}

