\name{my.axis}
\alias{my.axis}
\title{Add a pretty axis}
\usage{
  my.axis(side, shade_years = NULL, ylab = "", yticks = NA)
}
\arguments{
  \item{side}{Number indicating the side to add an axis (as
  in the number passed to \code{axis(side = ...)}}

  \item{shade_years}{An optional numerical vector of length
  two giving the minimum and maximum years over which to
  add a light grey shading.}

  \item{ylab}{Y axis label}

  \item{yticks}{Logical: should y-axis ticks be added?}
}
\description{
  Add a pretty axis
}

