% Generated by roxygen2 (4.0.0): do not edit by hand
\name{optim_thermal}
\alias{optim_thermal}
\title{Optimize to find optimal max productivity Ricker a}
\usage{
optim_thermal(optim_temp, width_param, desired_area)
}
\arguments{
  \item{optim_temp}{The optimum temperature as a numeric
  value}

  \item{width_param}{The width parameter as a numeric
  value}

  \item{desired_area}{The desired area as a numeric value}
}
\description{
Optimize to find optimal max productivity Ricker a
}

