\name{plot_cons_plans}
\alias{plot_cons_plans}
\title{Plot conservation plans in mean-variance space}
\usage{
  plot_cons_plans(plans_mv, plans_name, cols, xlim = NULL,
    ylim = NULL, add_pts = TRUE, add_all_efs = FALSE,
    x_axis = TRUE, y_axis = TRUE, add_legend = TRUE,
    w_show = "all",
    xlab = "Variance of\n    generation-to-generation\nrate of change",
    ylab = "Mean of\n    generation-to-generation\nrate of change")
}
\description{
  Plot conservation plans in mean-variance space
}

