\name{plot_cons_plans_ts}
\alias{plot_cons_plans_ts}
\title{Plot time series of abundance from conservation plans}
\usage{
  plot_cons_plans_ts(x, burn = 1:30)
}
\arguments{
  \item{x}{A list of output from \code{\link{meta_sim}}}
}
\description{
  Plot time series of abundance from conservation plans
}

