\name{plot_cons_plans_ts_panel}
\alias{plot_cons_plans_ts_panel}
\title{Plot conservation plan abundance time series panel}
\usage{
  plot_cons_plans_ts_panel(x)
}
\arguments{
  \item{x}{A \code{\link{meta_sim}} output object.}
}
\description{
  Plot conservation plan abundance time series panel
}

