\name{thermal_area}
\alias{thermal_area}
\title{Return desired squared deviation between desired area and actual
area under a curve}
\usage{
  thermal_area(max_a, desired_area, optim_temp,
    width_param, lower = -5, upper = 40)
}
\description{
  Return desired squared deviation between desired area and
  actual area under a curve
}

