% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_port_vals.R
\name{VaR}
\alias{VaR}
\title{Value at Risk}
\usage{
VaR(x, probs = 0.05)
}
\arguments{
\item{x}{A numeric vector}

\item{probs}{The probability cutoff to pass to the value at risk.}
}
\description{
Get the value at risk.
}
