% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cons_plans.R
\name{custom_bw}
\alias{custom_bw}
\title{Custom bandwidth}
\usage{
custom_bw(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Based on \code{bandwidth.nrd} from \pkg{MASS}. This version takes the
absolute value of \code{var} to avoid errors.
}
