% Generated by roxygen2 (4.0.0): do not edit by hand
\name{est_beta_params}
\alias{est_beta_params}
\title{Get beta parameters from mean and variance}
\usage{
est_beta_params(mu, var)
}
\arguments{
  \item{mu}{Mean}

  \item{var}{Variance}
}
\description{
Get beta parameters from mean and variance
}

