% Generated by roxygen2 (4.0.1): do not edit by hand
\name{get_quantile_contour}
\alias{get_quantile_contour}
\title{Get quantile contour}
\usage{
get_quantile_contour(x, alpha = 0.8)
}
\arguments{
  \item{x}{Output from \code{\link[MASS]{kde2d}}.}

  \item{alpha}{The quantile level.}
}
\description{
Get quantile contour
}

