% Generated by roxygen2 (4.0.0): do not edit by hand
\name{my.axis}
\alias{my.axis}
\title{Add a pretty axis}
\usage{
my.axis(side, shade_years = NULL, ylab = "", yticks = NA)
}
\arguments{
  \item{side}{Number indicating the side to add an axis (as
  in the side number passed to
  \code{\link[graphics]{axis}}).}

  \item{shade_years}{An optional numerical vector of length
  two giving the minimum and maximum years over which to
  add a light grey shading.}

  \item{ylab}{Y axis label}

  \item{yticks}{Logical: should y-axis ticks be added?}
}
\description{
Add a pretty axis
}

