\name{plot_cons_plans}
\alias{plot_cons_plans}
\title{Plot conservation plans in mean-variance space}
\usage{
  plot_cons_plans(plans_mv, plans_name, cols, xlim = NULL,
    ylim = NULL, add_pts = TRUE, add_all_efs = FALSE,
    w_show = "all")
}
\description{
  Plot conservation plans in mean-variance space
}

