\name{plot_sp_A_ts}
\alias{plot_sp_A_ts}
\title{TODO document plot_sp_A_ts}
\usage{
  plot_sp_A_ts(X, ylim, x_axis = TRUE, y_axis = TRUE,
    rate = FALSE, lwd = 1.7, y_axis_ticks = NULL,
    start_new_plots = 1, labels = NULL, burn = 30, cols,
    ...)
}
\arguments{
  \item{X}{object to plot}

  \item{ylim}{y limits}

  \item{x_axis}{TRUE}

  \item{y_axis}{TRUE}

  \item{rate}{FALSE}

  \item{lwd}{1.7}

  \item{y_axis_ticks}{NULL}

  \item{start_new_plots}{1}

  \item{labels}{NULL}

  \item{burn}{30}

  \item{cols}{Colours}

  \item{...}{anything else to pass to
  \code{plot.default()}}
}
\description{
  TODO document plot_sp_A_ts
}

